Rename to make diff less noisy.
